+++ src/RenderEngineManager.cc
@@ -83,7 +83,7 @@ class ignition::rendering::RenderEngineManagerPrivate
 
   /// \brief Plugin loader for managing render engine plugin libraries
   public: ignition::plugin::Loader pluginLoader;
-  
+
   /// \brief Environment variable which holds paths to look for plugins
   public: std::string pluginPathEnv = "IGN_RENDERING_PLUGIN_PATH";
 
@@ -377,7 +377,7 @@ bool RenderEngineManagerPrivate::LoadEnginePlugin(
 
   ignition::common::SystemPaths systemPaths;
   systemPaths.SetPluginPathEnv(this->pluginPathEnv);
-  
+
   // Add default install folder.
   systemPaths.AddPluginPaths(std::string(IGN_RENDERING_PLUGIN_PATH));
   // Add extra search path.